torchvideoclf/requirements.txt
aiohttp==3.7.4.post0
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bidict==0.21.2
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
contextvars==2.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.7
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==1.1.2
Flask-SocketIO==4.3.1
future==0.18.2
grpcio==1.36.1
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
immutables==0.15
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.1.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.3
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.17
ptyprocess==0.7.0
pycparser==2.20
pycuda==2020.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytools==2020.4.3
pytz==2021.1
pyzmq==22.0.3
regex==2021.3.17
requests==2.25.1
sacremoses==0.0.43
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
terminado==0.9.3
testpath==0.4.4
tokenizers==0.10.1
torch==1.7.0
torchaudio==0.6.0a0+f17ae39
torchvision==0.8.1
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
transformers==4.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
youtube-dl==2021.3.14
zipp==3.4.1